Add files via upload
New script that includes all the processes already used in visor_toponomastics.R plus new functions and examples of exploratory analysis of the data. 

Requires libraries: dplyr for easier dataframes manipulation and sf for shapefiles

Description of the script:

DATA LOAD
The data is now loaded automatically from the script (zip files unzipped and stored on user's folder)

DATA SELECTION
Selects the dataframes that are relevant for toponymic research from the data loaded

DATA CLEANING
Groups the data from shapefiles in a dataframe and processes the most relevant variables to create columns that contain the most relevant values

FUNCTIONS 
data_search
Searches a given theme or toponym. It allows searches for a specific entity type (point or polygon).

unique_toponym
Processes a given search to avoid reduplication of toponymic expressions when they originally belong to the same place (E.g. a place name of a particular village which is also used to name a whole parish or a municipality).

map_galiza
creates a map plotting a given theme with all occurrences found for the area of the data used in the study (Galiza)

map_galiza2
creates a map comparing two themes 

list_toponimos
creates a list of toponyms for a given theme

full_report
Searches a theme in the database, creates a list of place-names that contain the them and plots a map with the results. It allows entity_type specification (optional)

DATA SELECTION

How to use the functions (examples)